refactor(mcp): split mcp.rs into mcp/ directory with per-domain tool submodules
Promotes apps/things3-cli/src/mcp.rs (4,880 lines) to a directory module.
Handler methods move to tools/{tasks,projects,areas,tags,export,admin,prompts,resources}.rs
grouped by domain; mod.rs retains the struct, error types, dispatch, and server setup.
No behavior change — all 305+ tests pass.
