+++ apps/things3-cli/src/mcp/mod.rs
@@ -2,23 +2,21 @@
 
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::str::FromStr;
 use std::sync::Arc;
 #[cfg(target_os = "macos")]
 use things3_core::AppleScriptBackend;
 use things3_core::{
-    models::ThingsId, BackupManager, DataExporter, DeleteChildHandling, ExportData, ExportFormat,
-    McpServerConfig, MutationBackend, PerformanceMonitor, SqlxBackend, ThingsCache, ThingsConfig,
-    ThingsDatabase, ThingsError,
+    BackupManager, DataExporter, McpServerConfig, MutationBackend, PerformanceMonitor, SqlxBackend,
+    ThingsCache, ThingsConfig, ThingsDatabase, ThingsError,
 };
 use thiserror::Error;
 use tokio::sync::Mutex;
-use tracing::warn;
 
 pub mod io_wrapper;
 pub mod middleware;
 // pub mod performance_tests; // Temporarily disabled due to API changes
 pub mod test_harness;
+mod tools;
 
 use io_wrapper::{McpIo, StdIo};
 use middleware::{MiddlewareChain, MiddlewareConfig};
@@ -2211,1855 +2209,13 @@ impl ThingsMcpServer {
         result
     }
 
-    async fn handle_get_inbox(&self, args: Value) -> McpResult<CallToolResult> {
-        let limit = args
-            .get("limit")
-            .and_then(serde_json::Value::as_u64)
-            .map(|v| usize::try_from(v).unwrap_or(usize::MAX));
-
-        let tasks = self
-            .db
-            .get_inbox(limit)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_inbox", e))?;
-
-        let json = serde_json::to_string_pretty(&tasks)
-            .map_err(|e| McpError::serialization_failed("get_inbox serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_today(&self, args: Value) -> McpResult<CallToolResult> {
-        let limit = args
-            .get("limit")
-            .and_then(serde_json::Value::as_u64)
-            .map(|v| usize::try_from(v).unwrap_or(usize::MAX));
-
-        let tasks = self.db.get_today(limit).await.map_err(|e| {
-            // Include the actual error message for debugging
-            McpError::database_operation_failed(
-                "get_today",
-                things3_core::ThingsError::unknown(format!("Failed to get today's tasks: {}", e)),
-            )
-        })?;
-
-        let json = serde_json::to_string_pretty(&tasks)
-            .map_err(|e| McpError::serialization_failed("get_today serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_projects(&self, args: Value) -> McpResult<CallToolResult> {
-        let _area_uuid = args
-            .get("area_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok());
-
-        let projects = self
-            .db
-            .get_projects(None)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_projects", e))?;
-
-        let json = serde_json::to_string_pretty(&projects)
-            .map_err(|e| McpError::serialization_failed("get_projects serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_areas(&self, _args: Value) -> McpResult<CallToolResult> {
-        let areas = self
-            .db
-            .get_areas()
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_areas", e))?;
-
-        let json = serde_json::to_string_pretty(&areas)
-            .map_err(|e| McpError::serialization_failed("get_areas serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_search_tasks(&self, args: Value) -> McpResult<CallToolResult> {
-        let query = args
-            .get("query")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("query"))?;
-
-        let _limit = args
-            .get("limit")
-            .and_then(serde_json::Value::as_u64)
-            .map(|v| usize::try_from(v).unwrap_or(usize::MAX));
-
-        let tasks = self
-            .db
-            .search_tasks(query)
-            .await
-            .map_err(|e| McpError::database_operation_failed("search_tasks", e))?;
-
-        let json = serde_json::to_string_pretty(&tasks)
-            .map_err(|e| McpError::serialization_failed("search_tasks serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_logbook_search(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse all optional parameters
-        let search_text = args
-            .get("search_text")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let from_date = args
-            .get("from_date")
-            .and_then(|v| v.as_str())
-            .and_then(|s| chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").ok());
-
-        let to_date = args
-            .get("to_date")
-            .and_then(|v| v.as_str())
-            .and_then(|s| chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").ok());
-
-        let project_uuid: Option<ThingsId> = args
-            .get("project_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok());
-
-        let area_uuid: Option<ThingsId> = args
-            .get("area_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok());
-
-        let tags = args.get("tags").and_then(|v| v.as_array()).map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                .collect::<Vec<String>>()
-        });
-
-        let limit = args.get("limit").and_then(|v| v.as_u64()).map(|v| v as u32);
-        let offset = args
-            .get("offset")
-            .and_then(|v| v.as_u64())
-            .map(|v| v as u32);
-
-        // Call database method
-        let tasks = self
-            .db
-            .search_logbook(
-                search_text,
-                from_date,
-                to_date,
-                project_uuid,
-                area_uuid,
-                tags,
-                limit,
-                offset,
-            )
-            .await
-            .map_err(|e| McpError::database_operation_failed("logbook_search", e))?;
-
-        // Serialize results
-        let json = serde_json::to_string_pretty(&tasks)
-            .map_err(|e| McpError::serialization_failed("logbook_search serialization", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text { text: json }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_create_task(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse request from JSON
-        let request: things3_core::CreateTaskRequest =
-            serde_json::from_value(args).map_err(|e| {
-                McpError::invalid_parameter(
-                    "request",
-                    format!("Failed to parse create task request: {e}"),
-                )
-            })?;
-
-        // Create task
-        let uuid = self
-            .mutations
-            .create_task(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("create_task", e))?;
-
-        // Return created task UUID
-        let response = serde_json::json!({
-            "uuid": uuid,
-            "message": "Task created successfully"
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("create_task response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_update_task(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse request from JSON
-        let request: things3_core::UpdateTaskRequest =
-            serde_json::from_value(args).map_err(|e| {
-                McpError::invalid_parameter(
-                    "request",
-                    format!("Failed to parse update task request: {e}"),
-                )
-            })?;
-
-        // Update task
-        self.mutations
-            .update_task(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("update_task", e))?;
-
-        // Return success
-        let response = serde_json::json!({
-            "message": "Task updated successfully"
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("update_task response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_complete_task(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        self.mutations
-            .complete_task(&id)
-            .await
-            .map_err(|e| McpError::database_operation_failed("complete_task", e))?;
-
-        let response = serde_json::json!({
-            "message": "Task completed successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("complete_task response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_uncomplete_task(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        self.mutations
-            .uncomplete_task(&id)
-            .await
-            .map_err(|e| McpError::database_operation_failed("uncomplete_task", e))?;
-
-        let response = serde_json::json!({
-            "message": "Task marked as incomplete successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("uncomplete_task response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_delete_task(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        let child_handling_str = args
-            .get("child_handling")
-            .and_then(|v| v.as_str())
-            .unwrap_or("error");
-
-        let child_handling = match child_handling_str {
-            "cascade" => DeleteChildHandling::Cascade,
-            "orphan" => DeleteChildHandling::Orphan,
-            _ => DeleteChildHandling::Error,
-        };
-
-        self.mutations
-            .delete_task(&id, child_handling)
-            .await
-            .map_err(|e| McpError::database_operation_failed("delete_task", e))?;
-
-        let response = serde_json::json!({
-            "message": "Task deleted successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("delete_task response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
     // ============================================================================
     // Bulk Operation Handlers
     // ============================================================================
 
-    async fn handle_bulk_move(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse task UUIDs
-        let task_uuid_strs: Vec<String> = args
-            .get("task_uuids")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| McpError::invalid_parameter("task_uuids", "Array of UUIDs is required"))?
-            .iter()
-            .filter_map(|v| v.as_str().map(String::from))
-            .collect();
-
-        let task_uuids: Vec<ThingsId> = task_uuid_strs
-            .iter()
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("task_uuids", format!("Invalid ID: {e}"))
-                })
-            })
-            .collect::<McpResult<Vec<_>>>()?;
-
-        // Parse optional project_uuid
-        let project_uuid: Option<ThingsId> = args
-            .get("project_uuid")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("project_uuid", format!("Invalid ID: {e}"))
-                })
-            })
-            .transpose()?;
-
-        // Parse optional area_uuid
-        let area_uuid: Option<ThingsId> = args
-            .get("area_uuid")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("area_uuid", format!("Invalid ID: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let request = things3_core::models::BulkMoveRequest {
-            task_uuids,
-            project_uuid,
-            area_uuid,
-        };
-
-        let result = self
-            .mutations
-            .bulk_move(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("bulk_move", e))?;
-
-        let response = serde_json::json!({
-            "success": result.success,
-            "processed_count": result.processed_count,
-            "message": result.message
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("bulk_move response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_bulk_update_dates(&self, args: Value) -> McpResult<CallToolResult> {
-        use chrono::NaiveDate;
-
-        // Parse task UUIDs
-        let task_uuid_strs: Vec<String> = args
-            .get("task_uuids")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| McpError::invalid_parameter("task_uuids", "Array of UUIDs is required"))?
-            .iter()
-            .filter_map(|v| v.as_str().map(String::from))
-            .collect();
-
-        let task_uuids: Vec<ThingsId> = task_uuid_strs
-            .iter()
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("task_uuids", format!("Invalid ID: {e}"))
-                })
-            })
-            .collect::<McpResult<Vec<_>>>()?;
-
-        // Parse optional dates
-        let start_date = args
-            .get("start_date")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("start_date", format!("Invalid date format: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let deadline = args
-            .get("deadline")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("deadline", format!("Invalid date format: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let clear_start_date = args
-            .get("clear_start_date")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let clear_deadline = args
-            .get("clear_deadline")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let request = things3_core::models::BulkUpdateDatesRequest {
-            task_uuids,
-            start_date,
-            deadline,
-            clear_start_date,
-            clear_deadline,
-        };
-
-        let result = self
-            .mutations
-            .bulk_update_dates(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("bulk_update_dates", e))?;
-
-        let response = serde_json::json!({
-            "success": result.success,
-            "processed_count": result.processed_count,
-            "message": result.message
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("bulk_update_dates response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_bulk_complete(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse task UUIDs
-        let task_uuid_strs: Vec<String> = args
-            .get("task_uuids")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| McpError::invalid_parameter("task_uuids", "Array of UUIDs is required"))?
-            .iter()
-            .filter_map(|v| v.as_str().map(String::from))
-            .collect();
-
-        let task_uuids: Vec<ThingsId> = task_uuid_strs
-            .iter()
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("task_uuids", format!("Invalid ID: {e}"))
-                })
-            })
-            .collect::<McpResult<Vec<_>>>()?;
-
-        let request = things3_core::models::BulkCompleteRequest { task_uuids };
-
-        let result = self
-            .mutations
-            .bulk_complete(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("bulk_complete", e))?;
-
-        let response = serde_json::json!({
-            "success": result.success,
-            "processed_count": result.processed_count,
-            "message": result.message
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("bulk_complete response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_bulk_delete(&self, args: Value) -> McpResult<CallToolResult> {
-        // Parse task UUIDs
-        let task_uuid_strs: Vec<String> = args
-            .get("task_uuids")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| McpError::invalid_parameter("task_uuids", "Array of UUIDs is required"))?
-            .iter()
-            .filter_map(|v| v.as_str().map(String::from))
-            .collect();
-
-        let task_uuids: Vec<ThingsId> = task_uuid_strs
-            .iter()
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("task_uuids", format!("Invalid ID: {e}"))
-                })
-            })
-            .collect::<McpResult<Vec<_>>>()?;
-
-        let request = things3_core::models::BulkDeleteRequest { task_uuids };
-
-        let result = self
-            .mutations
-            .bulk_delete(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("bulk_delete", e))?;
-
-        let response = serde_json::json!({
-            "success": result.success,
-            "processed_count": result.processed_count,
-            "message": result.message
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("bulk_delete response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_create_project(&self, args: Value) -> McpResult<CallToolResult> {
-        let title = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("title", "Project title is required"))?
-            .to_string();
-
-        let notes = args.get("notes").and_then(|v| v.as_str()).map(String::from);
-
-        let area_uuid: Option<ThingsId> = args
-            .get("area_uuid")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("area_uuid", format!("Invalid ID: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let start_date = args
-            .get("start_date")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("start_date", format!("Invalid date: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let deadline = args
-            .get("deadline")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("deadline", format!("Invalid date: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let tags = args.get("tags").and_then(|v| v.as_array()).map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(String::from))
-                .collect::<Vec<_>>()
-        });
-
-        let request = things3_core::models::CreateProjectRequest {
-            title,
-            notes,
-            area_uuid,
-            start_date,
-            deadline,
-            tags,
-        };
-
-        let id = self
-            .mutations
-            .create_project(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("create_project", e))?;
-
-        let response = serde_json::json!({
-            "message": "Project created successfully",
-            "uuid": id.to_string()
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("create_project response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_update_project(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        let title = args.get("title").and_then(|v| v.as_str()).map(String::from);
-        let notes = args.get("notes").and_then(|v| v.as_str()).map(String::from);
-
-        let area_uuid: Option<ThingsId> = args
-            .get("area_uuid")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                ThingsId::from_str(s).map_err(|e| {
-                    McpError::invalid_parameter("area_uuid", format!("Invalid ID: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let start_date = args
-            .get("start_date")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("start_date", format!("Invalid date: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let deadline = args
-            .get("deadline")
-            .and_then(|v| v.as_str())
-            .map(|s| {
-                chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").map_err(|e| {
-                    McpError::invalid_parameter("deadline", format!("Invalid date: {e}"))
-                })
-            })
-            .transpose()?;
-
-        let tags = args.get("tags").and_then(|v| v.as_array()).map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(String::from))
-                .collect::<Vec<_>>()
-        });
-
-        let request = things3_core::models::UpdateProjectRequest {
-            uuid: id,
-            title,
-            notes,
-            area_uuid,
-            start_date,
-            deadline,
-            tags,
-        };
-
-        self.mutations
-            .update_project(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("update_project", e))?;
-
-        let response = serde_json::json!({
-            "message": "Project updated successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("update_project response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_complete_project(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        let child_handling_str = args
-            .get("child_handling")
-            .and_then(|v| v.as_str())
-            .unwrap_or("error");
-
-        let child_handling = match child_handling_str {
-            "cascade" => things3_core::models::ProjectChildHandling::Cascade,
-            "orphan" => things3_core::models::ProjectChildHandling::Orphan,
-            _ => things3_core::models::ProjectChildHandling::Error,
-        };
-
-        self.mutations
-            .complete_project(&id, child_handling)
-            .await
-            .map_err(|e| McpError::database_operation_failed("complete_project", e))?;
-
-        let response = serde_json::json!({
-            "message": "Project completed successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("complete_project response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_delete_project(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        let child_handling_str = args
-            .get("child_handling")
-            .and_then(|v| v.as_str())
-            .unwrap_or("error");
-
-        let child_handling = match child_handling_str {
-            "cascade" => things3_core::models::ProjectChildHandling::Cascade,
-            "orphan" => things3_core::models::ProjectChildHandling::Orphan,
-            _ => things3_core::models::ProjectChildHandling::Error,
-        };
-
-        self.mutations
-            .delete_project(&id, child_handling)
-            .await
-            .map_err(|e| McpError::database_operation_failed("delete_project", e))?;
-
-        let response = serde_json::json!({
-            "message": "Project deleted successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("delete_project response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_create_area(&self, args: Value) -> McpResult<CallToolResult> {
-        let title = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("title", "Area title is required"))?
-            .to_string();
-
-        let request = things3_core::models::CreateAreaRequest { title };
-
-        let uuid = self
-            .mutations
-            .create_area(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("create_area", e))?;
-
-        let response = serde_json::json!({
-            "message": "Area created successfully",
-            "uuid": uuid.to_string()
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("create_area response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_update_area(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        let title = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("title", "Title is required"))?
-            .to_string();
-
-        let request = things3_core::models::UpdateAreaRequest { uuid: id, title };
-
-        self.mutations
-            .update_area(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("update_area", e))?;
-
-        let response = serde_json::json!({
-            "message": "Area updated successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("update_area response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_delete_area(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid_str = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("uuid", "UUID is required"))?;
-
-        let id = ThingsId::from_str(uuid_str)
-            .map_err(|e| McpError::invalid_parameter("uuid", format!("Invalid ID: {e}")))?;
-
-        self.mutations
-            .delete_area(&id)
-            .await
-            .map_err(|e| McpError::database_operation_failed("delete_area", e))?;
-
-        let response = serde_json::json!({
-            "message": "Area deleted successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("delete_area response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_productivity_metrics(&self, args: Value) -> McpResult<CallToolResult> {
-        let days = usize::try_from(
-            args.get("days")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(7),
-        )
-        .unwrap_or(7);
-
-        // Get various metrics
-        let db = &self.db;
-        let inbox_tasks = db
-            .get_inbox(None)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_inbox for metrics", e))?;
-        let today_tasks = db
-            .get_today(None)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_today for metrics", e))?;
-        let projects = db
-            .get_projects(None)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_projects for metrics", e))?;
-        let areas = db
-            .get_areas()
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_areas for metrics", e))?;
-        let _ = db;
-
-        let metrics = serde_json::json!({
-            "period_days": days,
-            "inbox_tasks_count": inbox_tasks.len(),
-            "today_tasks_count": today_tasks.len(),
-            "projects_count": projects.len(),
-            "areas_count": areas.len(),
-            "completed_tasks": projects.iter().filter(|p| p.status == things3_core::TaskStatus::Completed).count(),
-            "incomplete_tasks": projects.iter().filter(|p| p.status == things3_core::TaskStatus::Incomplete).count(),
-            "timestamp": chrono::Utc::now()
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&metrics).map_err(|e| {
-                    McpError::serialization_failed("productivity_metrics serialization", e)
-                })?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_export_data(&self, args: Value) -> McpResult<CallToolResult> {
-        let format = args
-            .get("format")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("format"))?;
-        let data_type = args
-            .get("data_type")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("data_type"))?;
-        let output_path = args.get("output_path").and_then(|v| v.as_str());
-
-        if !matches!(data_type, "tasks" | "projects" | "areas" | "all") {
-            return Err(McpError::invalid_data_type(
-                data_type,
-                "tasks, projects, areas, all",
-            ));
-        }
-
-        if format == "csv" && data_type == "all" {
-            return Err(McpError::invalid_parameter(
-                "data_type",
-                "CSV format does not support data_type=all. Use tasks, projects, or areas individually.",
-            ));
-        }
-
-        let db = &self.db;
-        let need_tasks = matches!(data_type, "tasks" | "all");
-        let need_projects = matches!(data_type, "projects" | "all");
-        let need_areas = matches!(data_type, "areas" | "all");
-
-        let inbox = if need_tasks {
-            db.get_inbox(None)
-                .await
-                .map_err(|e| McpError::database_operation_failed("get_inbox for export", e))?
-        } else {
-            vec![]
-        };
-        let today = if need_tasks {
-            db.get_today(None)
-                .await
-                .map_err(|e| McpError::database_operation_failed("get_today for export", e))?
-        } else {
-            vec![]
-        };
-        let projects = if need_projects {
-            db.get_projects(None)
-                .await
-                .map_err(|e| McpError::database_operation_failed("get_projects for export", e))?
-        } else {
-            vec![]
-        };
-        let areas = if need_areas {
-            db.get_areas()
-                .await
-                .map_err(|e| McpError::database_operation_failed("get_areas for export", e))?
-        } else {
-            vec![]
-        };
-
-        let mut counts = serde_json::Map::new();
-        if need_tasks {
-            counts.insert("inbox".to_string(), inbox.len().into());
-            counts.insert("today".to_string(), today.len().into());
-        }
-        if need_projects {
-            counts.insert("projects".to_string(), projects.len().into());
-        }
-        if need_areas {
-            counts.insert("areas".to_string(), areas.len().into());
-        }
-
-        let formatted = match format {
-            "json" => {
-                let json_val = match data_type {
-                    "tasks" => serde_json::json!({ "inbox": &inbox, "today": &today }),
-                    "projects" => serde_json::json!({ "projects": &projects }),
-                    "areas" => serde_json::json!({ "areas": &areas }),
-                    _ => serde_json::json!({
-                        "inbox": &inbox,
-                        "today": &today,
-                        "projects": &projects,
-                        "areas": &areas
-                    }),
-                };
-                serde_json::to_string_pretty(&json_val)
-                    .map_err(|e| McpError::serialization_failed("export_data json", e))?
-            }
-            "csv" => {
-                let mut all_tasks = inbox;
-                all_tasks.extend(today);
-                let export_data = ExportData::new(all_tasks, projects, areas);
-                DataExporter::new_default()
-                    .export(&export_data, ExportFormat::Csv)
-                    .map_err(|e| McpError::configuration_error(e.to_string()))?
-            }
-            "markdown" => {
-                let mut all_tasks = inbox;
-                all_tasks.extend(today);
-                let export_data = ExportData::new(all_tasks, projects, areas);
-                DataExporter::new_default()
-                    .export(&export_data, ExportFormat::Markdown)
-                    .map_err(|e| McpError::configuration_error(e.to_string()))?
-            }
-            _ => return Err(McpError::invalid_format(format, "json, csv, markdown")),
-        };
-
-        if let Some(raw_path) = output_path {
-            let path = expand_tilde(raw_path)?;
-            let bytes = formatted.as_bytes();
-            std::fs::write(&path, bytes)
-                .map_err(|e| McpError::io_operation_failed("export_data write", e))?;
-            let confirmation = serde_json::json!({
-                "path": path.to_string_lossy().as_ref(),
-                "format": format,
-                "data_type": data_type,
-                "bytes_written": bytes.len(),
-                "counts": serde_json::Value::Object(counts)
-            });
-            Ok(CallToolResult {
-                content: vec![Content::Text {
-                    text: serde_json::to_string_pretty(&confirmation).map_err(|e| {
-                        McpError::serialization_failed("export_data confirmation", e)
-                    })?,
-                }],
-                is_error: false,
-            })
-        } else {
-            Ok(CallToolResult {
-                content: vec![Content::Text { text: formatted }],
-                is_error: false,
-            })
-        }
-    }
-
-    async fn handle_bulk_create_tasks(&self, args: Value) -> McpResult<CallToolResult> {
-        // Validate top-level shape before delegating so we keep the historical
-        // "missing tasks" error contract — `serde_json::from_value` would also
-        // reject this, but with a less actionable error.
-        args.get("tasks")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| McpError::missing_parameter("tasks"))?;
-
-        let request: things3_core::models::BulkCreateTasksRequest = serde_json::from_value(args)
-            .map_err(|e| {
-                McpError::invalid_parameter(
-                    "tasks",
-                    format!("Failed to parse bulk_create_tasks request: {e}"),
-                )
-            })?;
-
-        let result = self
-            .mutations
-            .bulk_create_tasks(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("bulk_create_tasks", e))?;
-
-        let response = serde_json::json!({
-            "success": result.success,
-            "processed_count": result.processed_count,
-            "message": result.message,
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("bulk_create_tasks response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_recent_tasks(&self, args: Value) -> McpResult<CallToolResult> {
-        let limit = args
-            .get("limit")
-            .and_then(serde_json::Value::as_u64)
-            .map(|v| usize::try_from(v).unwrap_or(usize::MAX));
-        let hours = i64::try_from(
-            args.get("hours")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(24),
-        )
-        .unwrap_or(24);
-
-        // For now, return inbox tasks as a proxy for recent tasks
-        // In a real implementation, this would query by creation/modification date
-        let tasks = self
-            .db
-            .get_inbox(limit)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_recent_tasks", e))?;
-
-        let response = serde_json::json!({
-            "message": "Recent tasks (using inbox as proxy)",
-            "hours_lookback": hours,
-            "tasks": tasks
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("get_recent_tasks response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_backup_database(&self, args: Value) -> McpResult<CallToolResult> {
-        let backup_dir = args
-            .get("backup_dir")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("backup_dir"))?;
-        let description = args.get("description").and_then(|v| v.as_str());
-
-        let backup_path = std::path::Path::new(backup_dir);
-        let metadata = self
-            .backup_manager
-            .lock()
-            .await
-            .create_backup(backup_path, description)
-            .map_err(|e| {
-                McpError::backup_operation_failed(
-                    "create_backup",
-                    things3_core::ThingsError::unknown(e.to_string()),
-                )
-            })?;
-
-        let response = serde_json::json!({
-            "message": "Backup created successfully",
-            "backup_path": metadata.backup_path,
-            "file_size": metadata.file_size,
-            "created_at": metadata.created_at
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("backup_database response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_restore_database(&self, args: Value) -> McpResult<CallToolResult> {
-        // Argument validation runs first so missing-parameter errors keep
-        // their existing semantics regardless of the safety gate below.
-        let backup_path = args
-            .get("backup_path")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("backup_path"))?;
-
-        // restore_database overwrites the live Things 3 SQLite file directly —
-        // CulturedCode-unsupported and the highest-corruption scenario rust-things3
-        // exposes. Decision per #126: keep the tool but gate it on (1) explicit
-        // --unsafe-direct-db opt-in AND (2) Things 3 must not be running.
-        if !self.unsafe_direct_db {
-            return Err(McpError::validation_error(
-                "restore_database is gated. Re-launch with --unsafe-direct-db \
-                 (or THINGS_UNSAFE_DIRECT_DB=1). It overwrites the live Things 3 \
-                 database directly — see https://culturedcode.com/things/support/articles/5510170/",
-            ));
-        }
-        if (self.process_check)() {
-            return Err(McpError::validation_error(
-                "restore_database refuses to run while Things 3 is open: \
-                 overwriting the database file under a running app corrupts it. \
-                 Quit Things 3 (Cmd-Q) and retry.",
-            ));
-        }
-
-        let backup_file = std::path::Path::new(backup_path);
-        self.backup_manager
-            .lock()
-            .await
-            .restore_backup(backup_file)
-            .map_err(|e| {
-                McpError::backup_operation_failed(
-                    "restore_backup",
-                    things3_core::ThingsError::unknown(e.to_string()),
-                )
-            })?;
-
-        let response = serde_json::json!({
-            "message": "Database restored successfully",
-            "backup_path": backup_path
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("restore_database response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_list_backups(&self, args: Value) -> McpResult<CallToolResult> {
-        let backup_dir = args
-            .get("backup_dir")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("backup_dir"))?;
-
-        let backup_path = std::path::Path::new(backup_dir);
-        let backups = self
-            .backup_manager
-            .lock()
-            .await
-            .list_backups(backup_path)
-            .map_err(|e| {
-                McpError::backup_operation_failed(
-                    "list_backups",
-                    things3_core::ThingsError::unknown(e.to_string()),
-                )
-            })?;
-
-        let response = serde_json::json!({
-            "backups": backups,
-            "count": backups.len()
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("list_backups response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_performance_stats(&self, _args: Value) -> McpResult<CallToolResult> {
-        let monitor = self.performance_monitor.lock().await;
-        let stats = monitor.get_all_stats();
-        let summary = monitor.get_summary();
-        drop(monitor);
-
-        let response = serde_json::json!({
-            "summary": summary,
-            "operation_stats": stats
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("performance_stats response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_system_metrics(&self, _args: Value) -> McpResult<CallToolResult> {
-        let metrics = self
-            .performance_monitor
-            .lock()
-            .await
-            .get_system_metrics()
-            .map_err(|e| {
-                McpError::performance_monitoring_failed(
-                    "get_system_metrics",
-                    things3_core::ThingsError::unknown(e.to_string()),
-                )
-            })?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&metrics)
-                    .map_err(|e| McpError::serialization_failed("system_metrics response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_cache_stats(&self, _args: Value) -> McpResult<CallToolResult> {
-        let stats = self.cache.lock().await.get_stats();
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&stats)
-                    .map_err(|e| McpError::serialization_failed("cache_stats response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
     // ========================================================================
     // TAG TOOL HANDLERS
     // ========================================================================
-
-    async fn handle_search_tags_tool(&self, args: Value) -> McpResult<CallToolResult> {
-        let query: String = args
-            .get("query")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("query", "Missing 'query' parameter"))?
-            .to_string();
-
-        let include_similar = args
-            .get("include_similar")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(true);
-
-        let min_similarity = args
-            .get("min_similarity")
-            .and_then(|v| v.as_f64())
-            .unwrap_or(0.7) as f32;
-
-        let tags = if include_similar {
-            self.db
-                .find_similar_tags(&query, min_similarity)
-                .await
-                .map_err(|e| McpError::database_operation_failed("search_tags", e))?
-                .into_iter()
-                .map(|tm| tm.tag)
-                .collect()
-        } else {
-            self.db
-                .search_tags(&query)
-                .await
-                .map_err(|e| McpError::database_operation_failed("search_tags", e))?
-        };
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&tags)
-                    .map_err(|e| McpError::serialization_failed("tags", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_tag_suggestions(&self, args: Value) -> McpResult<CallToolResult> {
-        let title: String = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("title", "Missing 'title' parameter"))?
-            .to_string();
-
-        use things3_core::database::tag_utils::normalize_tag_title;
-        let normalized = normalize_tag_title(&title);
-
-        // Check for exact match
-        let exact_match = self
-            .db
-            .find_tag_by_normalized_title(&normalized)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_tag_suggestions", e))?;
-
-        // Find similar tags
-        let similar_tags = self
-            .db
-            .find_similar_tags(&normalized, 0.7)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_tag_suggestions", e))?;
-
-        let recommendation = if exact_match.is_some() {
-            "use_existing"
-        } else if !similar_tags.is_empty() {
-            "consider_similar"
-        } else {
-            "create_new"
-        };
-
-        let response = serde_json::json!({
-            "exact_match": exact_match,
-            "similar_tags": similar_tags,
-            "recommendation": recommendation
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("tag_suggestions", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_popular_tags(&self, args: Value) -> McpResult<CallToolResult> {
-        let limit = args.get("limit").and_then(|v| v.as_u64()).unwrap_or(20) as usize;
-
-        let tags = self
-            .db
-            .get_popular_tags(limit)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_popular_tags", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&tags)
-                    .map_err(|e| McpError::serialization_failed("popular_tags", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_recent_tags(&self, args: Value) -> McpResult<CallToolResult> {
-        let limit = args.get("limit").and_then(|v| v.as_u64()).unwrap_or(20) as usize;
-
-        let tags = self
-            .db
-            .get_recent_tags(limit)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_recent_tags", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&tags)
-                    .map_err(|e| McpError::serialization_failed("recent_tags", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_create_tag(&self, args: Value) -> McpResult<CallToolResult> {
-        let title: String = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::invalid_parameter("title", "Missing 'title' parameter"))?
-            .to_string();
-
-        let shortcut: Option<String> = args
-            .get("shortcut")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let parent_uuid: Option<ThingsId> = args
-            .get("parent_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok());
-
-        let force = args.get("force").and_then(|v| v.as_bool()).unwrap_or(false);
-
-        let request = things3_core::models::CreateTagRequest {
-            title,
-            shortcut,
-            parent_uuid,
-        };
-
-        let result = match self
-            .mutations
-            .create_tag(request, force)
-            .await
-            .map_err(|e| McpError::database_operation_failed("create_tag", e))?
-        {
-            things3_core::models::TagCreationResult::Created { uuid, .. } => {
-                let message = if force {
-                    "Tag created successfully (duplicate check skipped)"
-                } else {
-                    "Tag created successfully"
-                };
-                serde_json::json!({
-                    "status": "created",
-                    "uuid": uuid,
-                    "message": message,
-                })
-            }
-            things3_core::models::TagCreationResult::Existing { tag, .. } => {
-                serde_json::json!({
-                    "status": "existing",
-                    "uuid": tag.uuid,
-                    "tag": tag,
-                    "message": "Tag already exists"
-                })
-            }
-            things3_core::models::TagCreationResult::SimilarFound {
-                similar_tags,
-                requested_title,
-            } => {
-                serde_json::json!({
-                    "status": "similar_found",
-                    "similar_tags": similar_tags,
-                    "requested_title": requested_title,
-                    "message": "Similar tags found. Use force=true to create anyway."
-                })
-            }
-        };
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&result)
-                    .map_err(|e| McpError::serialization_failed("create_tag_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_update_tag(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid: ThingsId = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("uuid", "Missing or invalid 'uuid' parameter")
-            })?;
-
-        let title: Option<String> = args
-            .get("title")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let shortcut: Option<String> = args
-            .get("shortcut")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let parent_uuid: Option<ThingsId> = args
-            .get("parent_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok());
-
-        let uuid_str = uuid.to_string();
-        let request = things3_core::models::UpdateTagRequest {
-            uuid,
-            title,
-            shortcut,
-            parent_uuid,
-        };
-
-        self.mutations
-            .update_tag(request)
-            .await
-            .map_err(|e| McpError::database_operation_failed("update_tag", e))?;
-
-        let response = serde_json::json!({
-            "message": "Tag updated successfully",
-            "uuid": uuid_str
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("update_tag_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_delete_tag(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid: ThingsId = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("uuid", "Missing or invalid 'uuid' parameter")
-            })?;
-
-        let remove_from_tasks = args
-            .get("remove_from_tasks")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        self.mutations
-            .delete_tag(&uuid, remove_from_tasks)
-            .await
-            .map_err(|e| McpError::database_operation_failed("delete_tag", e))?;
-
-        let response = serde_json::json!({
-            "message": "Tag deleted successfully",
-            "uuid": uuid
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("delete_tag_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_merge_tags(&self, args: Value) -> McpResult<CallToolResult> {
-        let source_uuid: ThingsId = args
-            .get("source_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter(
-                    "source_uuid",
-                    "Missing or invalid 'source_uuid' parameter",
-                )
-            })?;
-
-        let target_uuid: ThingsId = args
-            .get("target_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter(
-                    "target_uuid",
-                    "Missing or invalid 'target_uuid' parameter",
-                )
-            })?;
-
-        self.mutations
-            .merge_tags(&source_uuid, &target_uuid)
-            .await
-            .map_err(|e| McpError::database_operation_failed("merge_tags", e))?;
-
-        let response = serde_json::json!({
-            "message": "Tags merged successfully",
-            "source_uuid": source_uuid,
-            "target_uuid": target_uuid
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("merge_tags_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_add_tag_to_task(&self, args: Value) -> McpResult<CallToolResult> {
-        let task_uuid: ThingsId = args
-            .get("task_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("task_uuid", "Missing or invalid 'task_uuid' parameter")
-            })?;
-
-        let tag_title: String = args
-            .get("tag_title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("tag_title", "Missing 'tag_title' parameter")
-            })?
-            .to_string();
-
-        let result = self
-            .mutations
-            .add_tag_to_task(&task_uuid, &tag_title)
-            .await
-            .map_err(|e| McpError::database_operation_failed("add_tag_to_task", e))?;
-
-        let response = match result {
-            things3_core::models::TagAssignmentResult::Assigned { tag_uuid } => {
-                serde_json::json!({
-                    "status": "assigned",
-                    "tag_uuid": tag_uuid,
-                    "message": "Tag added to task successfully"
-                })
-            }
-            things3_core::models::TagAssignmentResult::Suggestions { similar_tags } => {
-                serde_json::json!({
-                    "status": "suggestions",
-                    "similar_tags": similar_tags,
-                    "message": "Similar tags found. Please confirm or use a different tag."
-                })
-            }
-        };
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("add_tag_to_task_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_remove_tag_from_task(&self, args: Value) -> McpResult<CallToolResult> {
-        let task_uuid: ThingsId = args
-            .get("task_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("task_uuid", "Missing or invalid 'task_uuid' parameter")
-            })?;
-
-        let tag_title: String = args
-            .get("tag_title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("tag_title", "Missing 'tag_title' parameter")
-            })?
-            .to_string();
-
-        self.mutations
-            .remove_tag_from_task(&task_uuid, &tag_title)
-            .await
-            .map_err(|e| McpError::database_operation_failed("remove_tag_from_task", e))?;
-
-        let response = serde_json::json!({
-            "message": "Tag removed from task successfully",
-            "task_uuid": task_uuid,
-            "tag_title": tag_title
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response).map_err(|e| {
-                    McpError::serialization_failed("remove_tag_from_task_response", e)
-                })?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_set_task_tags(&self, args: Value) -> McpResult<CallToolResult> {
-        let task_uuid: ThingsId = args
-            .get("task_uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("task_uuid", "Missing or invalid 'task_uuid' parameter")
-            })?;
-
-        let tag_titles: Vec<String> = args
-            .get("tag_titles")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("tag_titles", "Missing 'tag_titles' parameter")
-            })?
-            .iter()
-            .filter_map(|v| v.as_str().map(|s| s.to_string()))
-            .collect();
-
-        let suggestions = self
-            .mutations
-            .set_task_tags(&task_uuid, tag_titles.clone())
-            .await
-            .map_err(|e| McpError::database_operation_failed("set_task_tags", e))?;
-
-        let response = serde_json::json!({
-            "message": "Task tags updated successfully",
-            "task_uuid": task_uuid,
-            "tags": tag_titles,
-            "suggestions": suggestions
-        });
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&response)
-                    .map_err(|e| McpError::serialization_failed("set_task_tags_response", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_tag_statistics(&self, args: Value) -> McpResult<CallToolResult> {
-        let uuid: ThingsId = args
-            .get("uuid")
-            .and_then(|v| v.as_str())
-            .and_then(|s| ThingsId::from_str(s).ok())
-            .ok_or_else(|| {
-                McpError::invalid_parameter("uuid", "Missing or invalid 'uuid' parameter")
-            })?;
-
-        let stats = self
-            .db
-            .get_tag_statistics(&uuid)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_tag_statistics", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&stats)
-                    .map_err(|e| McpError::serialization_failed("tag_statistics", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_find_duplicate_tags(&self, args: Value) -> McpResult<CallToolResult> {
-        let min_similarity = args
-            .get("min_similarity")
-            .and_then(|v| v.as_f64())
-            .unwrap_or(0.85) as f32;
-
-        let duplicates = self
-            .db
-            .find_duplicate_tags(min_similarity)
-            .await
-            .map_err(|e| McpError::database_operation_failed("find_duplicate_tags", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&duplicates)
-                    .map_err(|e| McpError::serialization_failed("duplicate_tags", e))?,
-            }],
-            is_error: false,
-        })
-    }
-
-    async fn handle_get_tag_completions(&self, args: Value) -> McpResult<CallToolResult> {
-        let prefix: String = if let Some(v) = args.get("prefix").and_then(|v| v.as_str()) {
-            v.to_string()
-        } else if let Some(v) = args.get("partial_input").and_then(|v| v.as_str()) {
-            warn!("get_tag_completions: 'partial_input' is deprecated, use 'prefix' instead");
-            v.to_string()
-        } else {
-            return Err(McpError::invalid_parameter(
-                "prefix",
-                "Missing 'prefix' parameter",
-            ));
-        };
-
-        let limit = args.get("limit").and_then(|v| v.as_u64()).unwrap_or(10) as usize;
-
-        let completions = self
-            .db
-            .get_tag_completions(&prefix, limit)
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_tag_completions", e))?;
-
-        Ok(CallToolResult {
-            content: vec![Content::Text {
-                text: serde_json::to_string_pretty(&completions)
-                    .map_err(|e| McpError::serialization_failed("tag_completions", e))?,
-            }],
-            is_error: false,
-        })
-    }
 
     /// Get available MCP prompts
     fn get_available_prompts() -> Vec<Prompt> {
@@ -4195,353 +2351,6 @@ impl ThingsMcpServer {
         }
     }
 
-    /// Handle prompt request
-    async fn handle_prompt_request(&self, request: GetPromptRequest) -> McpResult<GetPromptResult> {
-        let prompt_name = &request.name;
-        let arguments = request.arguments.unwrap_or_default();
-
-        match prompt_name.as_str() {
-            "task_review" => self.handle_task_review_prompt(arguments).await,
-            "project_planning" => self.handle_project_planning_prompt(arguments).await,
-            "productivity_analysis" => self.handle_productivity_analysis_prompt(arguments).await,
-            "backup_strategy" => self.handle_backup_strategy_prompt(arguments).await,
-            _ => Err(McpError::prompt_not_found(prompt_name)),
-        }
-    }
-
-    /// Handle task review prompt
-    async fn handle_task_review_prompt(&self, args: Value) -> McpResult<GetPromptResult> {
-        let task_title = args
-            .get("task_title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("task_title"))?;
-        let task_notes = args.get("task_notes").and_then(|v| v.as_str());
-        let context = args.get("context").and_then(|v| v.as_str());
-
-        // Get current data for context
-        let db = &self.db;
-        let inbox_tasks = db
-            .get_inbox(Some(5))
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_inbox for task_review", e))?;
-        let today_tasks = db
-            .get_today(Some(5))
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_today for task_review", e))?;
-        let _ = db;
-
-        let prompt_text = format!(
-            "# Task Review: {}\n\n\
-            ## Current Task Details\n\
-            - **Title**: {}\n\
-            - **Notes**: {}\n\
-            - **Context**: {}\n\n\
-            ## Review Checklist\n\
-            Please review this task for:\n\
-            1. **Clarity**: Is the task title clear and actionable?\n\
-            2. **Completeness**: Does it have all necessary details?\n\
-            3. **Priority**: How urgent/important is this task?\n\
-            4. **Dependencies**: Are there any prerequisites?\n\
-            5. **Time Estimate**: How long should this take?\n\n\
-            ## Current Context\n\
-            - **Inbox Tasks**: {} tasks\n\
-            - **Today's Tasks**: {} tasks\n\n\
-            ## Recommendations\n\
-            Based on the current workload and task details, provide specific recommendations for:\n\
-            - Improving task clarity\n\
-            - Breaking down complex tasks\n\
-            - Setting appropriate deadlines\n\
-            - Managing dependencies\n\n\
-            ## Next Steps\n\
-            Suggest concrete next steps to move this task forward effectively.",
-            task_title,
-            task_title,
-            task_notes.unwrap_or("No notes provided"),
-            context.unwrap_or("No additional context"),
-            inbox_tasks.len(),
-            today_tasks.len()
-        );
-
-        Ok(GetPromptResult {
-            content: vec![Content::Text { text: prompt_text }],
-            is_error: false,
-        })
-    }
-
-    /// Handle project planning prompt
-    async fn handle_project_planning_prompt(&self, args: Value) -> McpResult<GetPromptResult> {
-        let project_title = args
-            .get("project_title")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("project_title"))?;
-        let project_description = args.get("project_description").and_then(|v| v.as_str());
-        let deadline = args.get("deadline").and_then(|v| v.as_str());
-        let complexity = args
-            .get("complexity")
-            .and_then(|v| v.as_str())
-            .unwrap_or("medium");
-
-        // Get current data for context
-        let db = &self.db;
-        let projects = db.get_projects(None).await.map_err(|e| {
-            McpError::database_operation_failed("get_projects for project_planning", e)
-        })?;
-        let areas = db.get_areas().await.map_err(|e| {
-            McpError::database_operation_failed("get_areas for project_planning", e)
-        })?;
-        let _ = db;
-
-        let prompt_text = format!(
-            "# Project Planning: {}\n\n\
-            ## Project Overview\n\
-            - **Title**: {}\n\
-            - **Description**: {}\n\
-            - **Deadline**: {}\n\
-            - **Complexity**: {}\n\n\
-            ## Planning Framework\n\
-            Please help plan this project by:\n\
-            1. **Breaking down** the project into manageable tasks\n\
-            2. **Estimating** time requirements for each task\n\
-            3. **Identifying** dependencies between tasks\n\
-            4. **Suggesting** milestones and checkpoints\n\
-            5. **Recommending** project organization (areas, tags, etc.)\n\n\
-            ## Current Context\n\
-            - **Existing Projects**: {} projects\n\
-            - **Available Areas**: {} areas\n\n\
-            ## Task Breakdown\n\
-            Create a detailed task list with:\n\
-            - Clear, actionable task titles\n\
-            - Estimated time for each task\n\
-            - Priority levels\n\
-            - Dependencies\n\
-            - Suggested deadlines\n\n\
-            ## Project Organization\n\
-            Suggest:\n\
-            - Appropriate area for this project\n\
-            - Useful tags for organization\n\
-            - Project structure and hierarchy\n\n\
-            ## Risk Assessment\n\
-            Identify potential challenges and mitigation strategies.\n\n\
-            ## Success Metrics\n\
-            Define how to measure project success and completion.",
-            project_title,
-            project_title,
-            project_description.unwrap_or("No description provided"),
-            deadline.unwrap_or("No deadline specified"),
-            complexity,
-            projects.len(),
-            areas.len()
-        );
-
-        Ok(GetPromptResult {
-            content: vec![Content::Text { text: prompt_text }],
-            is_error: false,
-        })
-    }
-
-    /// Handle productivity analysis prompt
-    async fn handle_productivity_analysis_prompt(&self, args: Value) -> McpResult<GetPromptResult> {
-        let time_period = args
-            .get("time_period")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("time_period"))?;
-        let focus_area = args
-            .get("focus_area")
-            .and_then(|v| v.as_str())
-            .unwrap_or("all");
-        let include_recommendations = args
-            .get("include_recommendations")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(true);
-
-        // Get current data for analysis
-        let db = &self.db;
-        let inbox_tasks = db.get_inbox(None).await.map_err(|e| {
-            McpError::database_operation_failed("get_inbox for productivity_analysis", e)
-        })?;
-        let today_tasks = db.get_today(None).await.map_err(|e| {
-            McpError::database_operation_failed("get_today for productivity_analysis", e)
-        })?;
-        let projects = db.get_projects(None).await.map_err(|e| {
-            McpError::database_operation_failed("get_projects for productivity_analysis", e)
-        })?;
-        let areas = db.get_areas().await.map_err(|e| {
-            McpError::database_operation_failed("get_areas for productivity_analysis", e)
-        })?;
-        let _ = db;
-
-        let completed_tasks = projects
-            .iter()
-            .filter(|p| p.status == things3_core::TaskStatus::Completed)
-            .count();
-        let incomplete_tasks = projects
-            .iter()
-            .filter(|p| p.status == things3_core::TaskStatus::Incomplete)
-            .count();
-
-        let prompt_text = format!(
-            "# Productivity Analysis - {}\n\n\
-            ## Analysis Period: {}\n\
-            ## Focus Area: {}\n\n\
-            ## Current Data Overview\n\
-            - **Inbox Tasks**: {} tasks\n\
-            - **Today's Tasks**: {} tasks\n\
-            - **Total Projects**: {} projects\n\
-            - **Areas**: {} areas\n\
-            - **Completed Tasks**: {} tasks\n\
-            - **Incomplete Tasks**: {} tasks\n\n\
-            ## Analysis Framework\n\
-            Please analyze productivity patterns focusing on:\n\n\
-            ### 1. Task Completion Patterns\n\
-            - Completion rates over the period\n\
-            - Task types that are completed vs. delayed\n\
-            - Time patterns in task completion\n\n\
-            ### 2. Workload Distribution\n\
-            - Balance between different areas/projects\n\
-            - Task complexity distribution\n\
-            - Deadline adherence patterns\n\n\
-            ### 3. Time Management\n\
-            - Task scheduling effectiveness\n\
-            - Inbox vs. scheduled task completion\n\
-            - Overdue task patterns\n\n\
-            ### 4. Project Progress\n\
-            - Project completion rates\n\
-            - Project complexity vs. completion time\n\
-            - Area-based productivity differences\n\n\
-            ## Key Insights\n\
-            Identify:\n\
-            - Peak productivity times\n\
-            - Most/least productive areas\n\
-            - Common bottlenecks\n\
-            - Success patterns\n\n\
-            ## Recommendations\n\
-            {}",
-            time_period,
-            time_period,
-            focus_area,
-            inbox_tasks.len(),
-            today_tasks.len(),
-            projects.len(),
-            areas.len(),
-            completed_tasks,
-            incomplete_tasks,
-            if include_recommendations {
-                "Provide specific, actionable recommendations for:\n\
-                - Improving task completion rates\n\
-                - Better time management\n\
-                - Workload balancing\n\
-                - Process optimization\n\
-                - Goal setting and tracking"
-            } else {
-                "Focus on analysis without recommendations"
-            }
-        );
-
-        Ok(GetPromptResult {
-            content: vec![Content::Text { text: prompt_text }],
-            is_error: false,
-        })
-    }
-
-    /// Handle backup strategy prompt
-    async fn handle_backup_strategy_prompt(&self, args: Value) -> McpResult<GetPromptResult> {
-        let data_volume = args
-            .get("data_volume")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("data_volume"))?;
-        let frequency = args
-            .get("frequency")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| McpError::missing_parameter("frequency"))?;
-        let retention_period = args
-            .get("retention_period")
-            .and_then(|v| v.as_str())
-            .unwrap_or("3_months");
-        let storage_preference = args
-            .get("storage_preference")
-            .and_then(|v| v.as_str())
-            .unwrap_or("hybrid");
-
-        // Get current data for context
-        let db = &self.db;
-        let projects = db.get_projects(None).await.map_err(|e| {
-            McpError::database_operation_failed("get_projects for backup_strategy", e)
-        })?;
-        let areas = db
-            .get_areas()
-            .await
-            .map_err(|e| McpError::database_operation_failed("get_areas for backup_strategy", e))?;
-        let _ = db;
-
-        let prompt_text = format!(
-            "# Backup Strategy Recommendation\n\n\
-            ## Requirements\n\
-            - **Data Volume**: {}\n\
-            - **Backup Frequency**: {}\n\
-            - **Retention Period**: {}\n\
-            - **Storage Preference**: {}\n\n\
-            ## Current Data Context\n\
-            - **Projects**: {} projects\n\
-            - **Areas**: {} areas\n\
-            - **Database Type**: SQLite (Things 3)\n\n\
-            ## Backup Strategy Analysis\n\n\
-            ### 1. Data Assessment\n\
-            Analyze the current data volume and growth patterns:\n\
-            - Database size estimation\n\
-            - Growth rate projections\n\
-            - Critical data identification\n\n\
-            ### 2. Backup Frequency Optimization\n\
-            For {} frequency backups:\n\
-            - Optimal timing considerations\n\
-            - Incremental vs. full backup strategy\n\
-            - Performance impact analysis\n\n\
-            ### 3. Storage Strategy\n\
-            For {} storage preference:\n\
-            - Local storage recommendations\n\
-            - Cloud storage options\n\
-            - Hybrid approach benefits\n\
-            - Cost considerations\n\n\
-            ### 4. Retention Policy\n\
-            For {} retention period:\n\
-            - Data lifecycle management\n\
-            - Compliance considerations\n\
-            - Storage optimization\n\n\
-            ## Recommended Implementation\n\
-            Provide specific recommendations for:\n\
-            - Backup tools and software\n\
-            - Storage locations and providers\n\
-            - Automation setup\n\
-            - Monitoring and alerting\n\
-            - Recovery procedures\n\n\
-            ## Risk Mitigation\n\
-            Address:\n\
-            - Data loss prevention\n\
-            - Backup verification\n\
-            - Disaster recovery planning\n\
-            - Security considerations\n\n\
-            ## Cost Analysis\n\
-            Estimate costs for:\n\
-            - Storage requirements\n\
-            - Backup software/tools\n\
-            - Cloud services\n\
-            - Maintenance overhead",
-            data_volume,
-            frequency,
-            retention_period,
-            storage_preference,
-            projects.len(),
-            areas.len(),
-            frequency,
-            storage_preference,
-            retention_period
-        );
-
-        Ok(GetPromptResult {
-            content: vec![Content::Text { text: prompt_text }],
-            is_error: false,
-        })
-    }
-
     /// Get available MCP resources
     fn get_available_resources() -> Vec<Resource> {
         vec![
@@ -4570,58 +2379,6 @@ impl ThingsMcpServer {
                 mime_type: Some("application/json".to_string()),
             },
         ]
-    }
-
-    /// Handle resource read request
-    async fn handle_resource_read(
-        &self,
-        request: ReadResourceRequest,
-    ) -> McpResult<ReadResourceResult> {
-        let uri = &request.uri;
-
-        let db = &self.db;
-        let data = match uri.as_str() {
-            "things://inbox" => {
-                let tasks = db.get_inbox(None).await.map_err(|e| {
-                    McpError::database_operation_failed("get_inbox for resource", e)
-                })?;
-                serde_json::to_string_pretty(&tasks).map_err(|e| {
-                    McpError::serialization_failed("inbox resource serialization", e)
-                })?
-            }
-            "things://projects" => {
-                let projects = db.get_projects(None).await.map_err(|e| {
-                    McpError::database_operation_failed("get_projects for resource", e)
-                })?;
-                serde_json::to_string_pretty(&projects).map_err(|e| {
-                    McpError::serialization_failed("projects resource serialization", e)
-                })?
-            }
-            "things://areas" => {
-                let areas = db.get_areas().await.map_err(|e| {
-                    McpError::database_operation_failed("get_areas for resource", e)
-                })?;
-                serde_json::to_string_pretty(&areas).map_err(|e| {
-                    McpError::serialization_failed("areas resource serialization", e)
-                })?
-            }
-            "things://today" => {
-                let tasks = db.get_today(None).await.map_err(|e| {
-                    McpError::database_operation_failed("get_today for resource", e)
-                })?;
-                let _ = db;
-                serde_json::to_string_pretty(&tasks).map_err(|e| {
-                    McpError::serialization_failed("today resource serialization", e)
-                })?
-            }
-            _ => {
-                return Err(McpError::resource_not_found(uri));
-            }
-        };
-
-        Ok(ReadResourceResult {
-            contents: vec![Content::Text { text: data }],
-        })
     }
 
     /// Handle a JSON-RPC request and return a JSON-RPC response
@@ -4793,7 +2550,7 @@ impl ThingsMcpServer {
     }
 }
 
-fn expand_tilde(path: &str) -> McpResult<std::path::PathBuf> {
+pub(in crate::mcp) fn expand_tilde(path: &str) -> McpResult<std::path::PathBuf> {
     if path == "~" || path.starts_with("~/") {
         let home = std::env::var("HOME").map_err(|_| {
             McpError::invalid_parameter(
