+++ apps/things3-cli/src/mcp/tools/mod.rs
@@ -1,0 +1,8 @@
+mod admin;
+mod areas;
+mod export;
+mod projects;
+mod prompts;
+mod resources;
+mod tags;
+mod tasks;
